Get rid of 'init', add suites
