+++ pkg/tools/security/test/provider.go
@@ -1,0 +1,35 @@
+// Copyright (c) 2020 Doc.ai and/or its affiliates.
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at:
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package securitytest
+
+import (
+	"context"
+	"crypto/tls"
+)
+
+type Provider struct {
+	GetTLSConfigFunc   func(ctx context.Context) (*tls.Config, error)
+	GetCertificateFunc func(ctx context.Context) (*tls.Certificate, error)
+}
+
+func (t *Provider) GetTLSConfig(ctx context.Context) (*tls.Config, error) {
+	return t.GetTLSConfigFunc(ctx)
+}
+
+func (t *Provider) GetCertificate(ctx context.Context) (*tls.Certificate, error) {
+	return t.GetCertificateFunc(ctx)
+}
